Refactor into blueprint
