+++ backend/extensions.py
@@ -1,0 +1,8 @@
+# extensions.py
+from flask_sqlalchemy import SQLAlchemy
+from flask_bcrypt import Bcrypt
+from flask_jwt_extended import JWTManager, create_access_token, jwt_required, get_jwt_identity, verify_jwt_in_request
+
+db = SQLAlchemy()
+bcrypt = Bcrypt()
+jwt = JWTManager()
